propagate error code and exit with negative errno
On failure, "1" was returned by the program. This is problematic when
checking for proper execution, since this value can also be returned by
the count at the end of the program.

Instead, we force all return code to be negative. In addition, we
propagate the error code when it was available from another command
(ioctl mainly). When it was not availabe, i.e. it is an error condition
within the program checks, we exit with an errno which seemed
appropriate for the condition.

Hopefully, this allows better analysis of the program's termination
state and location of the source of test failures.